Resolve padbook issues with GetNextPad
The issue was that the directory for a contact could exist, but if there was no pad.txt file in that dir then this function would break. Refactored the Padbook constructor to take either the full path to pad.txt, or take a path to /contacts and a username... working with all these paths can get confusing on which is being passed in, so having this constructor support both is helpful.